requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
Flask==1.1.2
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonpickle==1.4
MarkupSafe==1.1.1
numpy==1.18.2
opencv-python==4.2.0.34
Pillow==7.1.1
pybase64==1.0.1
requests==2.23.0
titlecase==0.12.0
urllib3==1.25.9
Werkzeug==1.0.1
zipp==3.1.0
